+++ test/functional-cloud/datastoresrp/resources/redis_test.go
@@ -25,57 +25,6 @@ import (
 	"github.com/radius-project/radius/test/validation"
 )
 
-func Test_Redis_Manual(t *testing.T) {
-	template := "testdata/datastoresrp-resources-redis-manual.bicep"
-	name := "dsrp-resources-redis-manual"
-	appNamespace := "default-dsrp-resources-redis-manual"
-
-	test := shared.NewRPTest(t, name, []shared.TestStep{
-		{
-			Executor: step.NewDeployExecutor(template, functional.GetMagpieImage()),
-			RPResources: &validation.RPResourceSet{
-				Resources: []validation.RPResource{
-					{
-						Name: name,
-						Type: validation.ApplicationsResource,
-					},
-					{
-						Name: "rds-app-ctnr",
-						Type: validation.ContainersResource,
-						App:  name,
-					},
-					{
-						Name: "rds-ctnr",
-						Type: validation.ContainersResource,
-						App:  name,
-					},
-					{
-						Name: "rds-rte",
-						Type: validation.HttpRoutesResource,
-						App:  name,
-					},
-					{
-						Name: "rds-rds",
-						Type: validation.RedisCachesResource,
-						App:  name,
-					},
-				},
-			},
-			K8sObjects: &validation.K8sObjectSet{
-				Namespaces: map[string][]validation.K8sObject{
-					appNamespace: {
-						validation.NewK8sPodForResource(name, "rds-app-ctnr"),
-						validation.NewK8sPodForResource(name, "rds-ctnr"),
-						validation.NewK8sServiceForResource(name, "rds-rte"),
-					},
-				},
-			},
-		},
-	})
-
-	test.Test(t)
-}
-
 func Test_Redis_Recipe(t *testing.T) {
 	template := "testdata/datastoresrp-resources-redis-recipe.bicep"
 	name := "dsrp-resources-redis-recipe"
