+++ test/functional/datastoresrp/resources/testdata/datastoresrp-resources-redis-manual.bicep
